# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "CadastroUsuarioRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "perfil": {
            "enum": [
              "servidor",
              "gestor",
              "administrador"
            ],
            "title": "Perfil",
            "type": "string"
          },
          "unidade_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Id"
          }
        },
        "required": [
          "nome",
          "email",
          "perfil"
        ],
        "title": "CadastroUsuarioRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListaUsuariosResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UsuarioResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ListaUsuariosResponse",
        "type": "object"
      },
...
      "MensagemResponse": {
        "properties": {
          "mensagem": {
            "title": "Mensagem",
            "type": "string"
          }
        },
        "required": [
          "mensagem"
        ],
        "title": "MensagemResponse",
        "type": "object"
      },
      "MeuPerfilResponse": {
        "description": "US 1.5 — dados cadastrais + histórico. `processos`/`documentos_assinados`\nficam sempre vazios neste change (as tabelas `processo`/`documento` são do\nÉpico 2/3); os campos de mensagem cobrem o estado vazio (Cen.2).",
        "properties": {
          "documentos_assinados": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Documentos Assinados",
            "type": "array"
          },
          "mensagem_documentos": {
            "default": "Nenhum documento assinado",
            "title": "Mensagem Documentos",
            "type": "string"
          },
          "mensagem_processos": {
            "default": "Nenhum processo registrado",
            "title": "Mensagem Processos",
            "type": "string"
          },
          "processos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Processos",
            "type": "array"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioResponse"
          }
        },
        "required": [
          "usuario"
        ],
        "title": "MeuPerfilResponse",
        "type": "object"
      },
...
      "TransferirUnidadeRequest": {
        "properties": {
          "unidade_id": {
            "title": "Unidade Id",
            "type": "string"
          }
        },
        "required": [
          "unidade_id"
        ],
        "title": "TransferirUnidadeRequest",
        "type": "object"
      },
...
      "UnidadesGeridasRequest": {
        "properties": {
          "unidade_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Unidade Ids",
            "type": "array"
          }
        },
        "required": [
          "unidade_ids"
        ],
        "title": "UnidadesGeridasRequest",
        "type": "object"
      },
      "UsuarioResponse": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "perfil": {
            "title": "Perfil",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unidade_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Id"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "perfil",
          "status",
          "unidade_id"
        ],
        "title": "UsuarioResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SETES.DOCS API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/usuarios/{usuario_id}/resetar-senha": {
      "post": {
        "description": "PRD US 1.10 — invalida a senha atual, envia link de redefinição, audita.",
        "operationId": "resetar_senha_admin_admin_usuarios__usuario_id__resetar_senha_post",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Usuario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resetar Senha Admin",
        "tags": [
          "usuarios"
        ]
      }
    },
...
    "/usuarios": {
      "get": {
        "description": "Administrador lista todos; Gestor só lista usuários das unidades que gerencia.",
        "operationId": "listar_usuarios_usuarios_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaUsuariosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Usuarios",
        "tags": [
          "usuarios"
        ]
      },
      "post": {
        "description": "PRD US 1.1 (Administrador) e US 1.2 (Gestor, restrito à própria unidade/perfil Servidor).",
        "operationId": "cadastrar_usuario_usuarios_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastroUsuarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cadastrar Usuario",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/me/perfil": {
      "get": {
        "description": "US 1.5 — sempre os dados do próprio token (`get_current_user`); não há\nparâmetro de ID manipulável, então não existe rota que permita ver o\nperfil de terceiros (US 1.5 Cen.2 / task 10.2).",
        "operationId": "meu_perfil_usuarios_me_perfil_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeuPerfilResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Meu Perfil",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/{usuario_id}/unidade": {
      "patch": {
        "description": "US 8.6 — vínculo 1:1 Servidor↔unidade. O PATCH sempre *substitui* o\nvínculo (nunca adiciona), então o cenário de \"vínculo duplo\" (Cen.2) é\nestruturalmente impossível: não existe operação que preserve o vínculo\nanterior e crie um segundo — `unidade_id` é uma FK escalar.",
        "operationId": "transferir_unidade_usuarios__usuario_id__unidade_patch",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Usuario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferirUnidadeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transferir Unidade",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/{usuario_id}/unidades-geridas": {
      "get": {
        "operationId": "obter_unidades_geridas_usuarios__usuario_id__unidades_geridas_get",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Usuario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Obter Unidades Geridas Usuarios  Usuario Id  Unidades Geridas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Obter Unidades Geridas",
        "tags": [
          "usuarios"
        ]
      },
      "put": {
        "description": "US 8.6b — conjunto de unidades geridas por um Gestor (N:N).",
        "operationId": "definir_unidades_geridas_usuarios__usuario_id__unidades_geridas_put",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Usuario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnidadesGeridasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Definir Unidades Geridas Usuarios  Usuario Id  Unidades Geridas Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Definir Unidades Geridas",
        "tags": [
          "usuarios"
        ]
      }
    }
  }
}